+++ example/wasm_decode.c
@@ -3,17 +3,15 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-// Optional: a free wrapper for JS
 EMSCRIPTEN_KEEPALIVE
 void pcc_free(void *ptr)
 {
   free(ptr);
 }
 
-// Just make sure pcc_decode is kept alive by the linker
 EMSCRIPTEN_KEEPALIVE
 void wasm_pcc_decode(const char     *rbuff,
-                     size_t          rsize,
+                     uint64_t        rsize,
                      float         **positions,
                      unsigned char **colors,
                      uint32_t       *num_points)
