+++ include/pcc/pcc.h
@@ -16,24 +16,24 @@ extern "C"
                              uint32_t             num_points,
                              const float          error,
                              char               **wbuff,
-                             size_t              *wsize);
+                             uint64_t            *wsize);
 
   PCC_EXPORT void pcc_decode(const char     *rbuff,
-                             size_t          rsize,
+                             uint64_t        rsize,
                              float         **positions,
                              unsigned char **colors,
                              uint32_t       *num_points);
 
   PCC_EXPORT void pcc_multiple_encode(const pcc_point_cloud_t *pcds,
-                                      size_t  num_pcds,
-                                      float   error,
-                                      char  **wbuff,
-                                      size_t *wsize);
+                                      uint64_t  num_pcds,
+                                      float     error,
+                                      char    **wbuff,
+                                      uint64_t *wsize);
 
   PCC_EXPORT void pcc_multiple_decode(const char         *rbuff,
-                                      size_t              rsize,
+                                      uint64_t            rsize,
                                       pcc_point_cloud_t **pcds_out,
-                                      size_t *num_pcds_out);
+                                      uint64_t *num_pcds_out);
 
 #ifdef __cplusplus
 } // extern "C"
